thêm shedlock

+++ backend/src/main/java/com/library/dto/response/TopBookProjection.java
@@ -6,6 +6,6 @@ public interface TopBookProjection {
     String getAuthor();
     String getCategory();
     String getImageUrl();
-    Long getBorrowCount();
+    Integer getBorrowCount();
 }
